Moved CompanyView to its own folder, and added SCSS module support
