+++ app/imports/ui/App.jsx
@@ -19,7 +19,7 @@ import ScriptEdit from './target-types/scripts/ScriptEdit';
 import PersonView from './target-types/person/PersonView';
 import PersonCreate from './target-types/person/PersonCreate';
 
-import CompanyView from './target-types/company/CompanyView';
+import CompanyView from './target-types/company/company-view/CompanyView';
 import CompanyCreate from './target-types/company/CompanyCreate';
 
 import DeviceView from './target-types/device/DeviceView';
