+++ app/imports/ui/target-types/company/CompanyCreateFormArray.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import { Form, Field, FieldArray } from 'formik';
-import { TextField, MaskedTextField, PrimaryButton, DefaultButton } from 'office-ui-fabric-react';
+import { Form } from 'formik';
+import { TextField, DefaultButton } from 'office-ui-fabric-react';
 
 export const CompanyCreateFormArray = ({ move, swap, push, insert, unshift, pop, form }) => (
 	<Form>
