+++ app/package.json
@@ -20,7 +20,7 @@
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
     "react-dropzone": "^7.0.1",
-    "react-redux": "^7.1.0",
+    "react-redux": "^7.1.1",
     "react-router": "^4.3.1",
     "react-router-dom": "^4.3.1",
     "react-table": "^6.10.0",
